separated views and functions into individual files
