+++ lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:ddsr/views/home.dart';
+import 'package:ddsr/views/second.dart';
 
 void main() => runApp(MyApp());
 
@@ -18,75 +20,3 @@ class MyApp extends StatelessWidget {
 }
 
 
-
-class BaseDrawer extends StatelessWidget{
-  String getCurrentRouteName(context) {
-    String currentRouteName;
-
-    Navigator.popUntil(context, (route) {
-      currentRouteName = route.settings.name;
-      return true;
-    });
-
-    return currentRouteName;
-  }
-  @override
-  Widget build(BuildContext context){
-    String currentRoute = getCurrentRouteName(context);
-    return Drawer(
-      child: ListView(
-        padding: EdgeInsets.zero,
-        children: <Widget>[
-          ListTile(
-            title: Text('First Page'),
-            onTap: (){
-              if(currentRoute == '/') return;
-              Navigator.pop(context);
-              Navigator.of(context).pushNamed('/');
-            }
-          ),
-          ListTile(
-            title: Text('Second Page'),
-            onTap: (){
-              if(currentRoute == '/second') return;
-              Navigator.pop(context);
-              Navigator.of(context).pushNamed('/second');
-            }
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-String getCurrentRouteName(context) {
-  String currentRouteName;
-
-  Navigator.popUntil(context, (route) {
-    currentRouteName = route.settings.name;
-    return true;
-  });
-  return currentRouteName;
-}
-
-class MyHome extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: new Text('asdf')),
-      body: Text('first page'),
-      drawer: BaseDrawer(),
-    );
-  }
-}
-
-class SecondPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title:  new Text('asdf')),
-      body: Text('second page'),
-      drawer: BaseDrawer(),
-    );
-  }
-}
